+++ rsync/config/__init__.py
@@ -1,0 +1,3 @@
+from .local_config import LocalConfig
+from .remote_config import RemoteConfig
+from .base_config import InvalidPath
