+++ rsync/config/base_config.py
